audio controller added
