+++ client/src/views/Home.tsx
@@ -15,6 +15,7 @@ import catchAsyncError from 'src/api/catchError';
 import {getClient} from 'src/api/client';
 import {useFetchPlaylist} from 'src/hooks/query';
 import {upldateNotification} from 'src/store/notification';
+import useAudioController from 'src/hooks/useAudioController';
 
 interface Props {}
 
@@ -23,6 +24,8 @@ const Home: FC<Props> = props => {
   const [selectedAudio, setSelectedAudio] = useState<AudioData>();
   const [showPlaylistModal, setShowPlaylistModal] = useState(false);
   const [showPlaylistForm, setShowPlaylistForm] = useState(false);
+
+  const {onAudioPress} = useAudioController();
 
   const {data} = useFetchPlaylist();
 
@@ -106,27 +109,11 @@ const Home: FC<Props> = props => {
   return (
     <View style={styles.container}>
       <LatestUploads
-        onAudioPress={async (item, data) => {
-          const lists: Track[] = data.map(item => {
-            return {
-              id: item.id,
-              title: item.title,
-              url: item.file,
-              artwork: item.poster || require('../assets/music.png'),
-              artist: item.owner.name,
-              genre: item.category,
-              isLiveStream: true,
-            };
-          });
-          await TrackPlayer.add([...lists]);
-          await TrackPlayer.play();
-        }}
+        onAudioPress={onAudioPress}
         onAudioLongPress={handleOnLongPress}
       />
       <RecommendedAudios
-        onAudioPress={item => {
-          console.log(item);
-        }}
+        onAudioPress={onAudioPress}
         onAudioLongPress={handleOnLongPress}
       />
       <OptionsModal
